+++ script/module/window.js
@@ -20,7 +20,7 @@ import {
 import {
     getFocusedID,
     getTargetBlockID,
-    getTargetHref,
+    getTargetInboxID,
 } from './../utils/dom.js';
 
 function open(id = getFocusedID(), urlParams = {}) {
@@ -53,7 +53,35 @@ function infocus(id = getFocusedID()) {
     });
 }
 
-async function middleClick(e, fn1, fn2 = null) {
+async function middleClick(e, fn1, fn2 = null, fn3 = null) {
+    // 收集箱
+    let inbox = getTargetInboxID(e.target);
+    if (inbox) {
+        for (const dock of ['leftDock', 'rightDock', 'topDock', 'bottomDock']) {
+            if (window.siyuan.layout[dock]
+                && window.siyuan.layout[dock].data
+                && window.siyuan.layout[dock].data.inbox
+                && window.siyuan.layout[dock].data.inbox.data
+                && window.siyuan.layout[dock].data.inbox.data[inbox]
+            ) {
+                if (fn3) {
+                    fn3(window.siyuan.layout[dock].data.inbox.data[inbox]);
+                    return;
+                }
+                else if (window.siyuan.layout[dock].data.inbox.data[inbox].shorthandURL) {
+                    // 新窗口打开收集箱超链接
+                    window.theme.openNewWindow(
+                        'browser',
+                        window.siyuan.layout[dock].data.inbox.data[inbox].shorthandURL,
+                        undefined,
+                        config.theme.window.open.windowParams,
+                    );
+                    return;
+                }
+            }
+        }
+    }
+
     let target = getTargetBlockID(e.target);
     if (target) {
         // 目标非空, 是 ID 或者链接
@@ -213,12 +241,7 @@ setTimeout(async () => {
                                                             config.theme.window.open.windowParams,
                                                             undefined,
                                                             undefined,
-                                                            async (win, event, level, message, line, sourceId) => {
-                                                                // 根据子窗口的控制台输出内容删除临时文件
-                                                                if (level === 0 && message === 'LOADED') {
-                                                                    setTimeout(async () => rm(template_path_absolute), 0);
-                                                                }
-                                                            },
+                                                            undefined,
                                                             async (win) => {
                                                                 // 窗口关闭时删除临时文件
                                                                 setTimeout(async () => rm(template_path_absolute), 0);
@@ -321,6 +344,48 @@ setTimeout(async () => {
                                                 config.theme.window.open.windowParams,
                                             );
                                         }
+                                    },
+                                    async inbox => {
+                                        // 收集箱
+                                        // 内容写入临时文件 => 打开编辑器 => 读取临时文件 => 删除临时文件
+                                        const filename = `${inbox.oId}.md`; // 临时文件文件名
+                                        // 临时文件绝对路径
+                                        const temp_file_path_absolute = `${config.theme.window.open.editor.path.temp.absolute}${filename}`;
+                                        // 临时文件相对路径(相对于思源工作空间根目录)
+                                        const temp_file_path_relative = `${config.theme.window.open.editor.path.temp.relative}${filename}`;
+                                        // console.log(temp_file_path_relative, temp_file_path_absolute);
+                                        // 写入临时文件
+                                        putFile(temp_file_path_relative, inbox.shorthandContent).then(r => {
+                                            if (r && r.code === 0) {
+                                                // 写入文件成功
+                                                window.theme.win = window.theme.openNewWindow(
+                                                    'editor',
+                                                    config.theme.window.open.editor.path.index,
+                                                    {
+                                                        mode: 'inbox',
+                                                        url: encodeURI(inbox.shorthandURL),
+                                                        path: encodeURI(temp_file_path_relative),
+                                                        lang: window.theme.languageMode,
+                                                        // theme: window.siyuan.config.appearance.mode,
+                                                        tabSize: window.siyuan.config.editor.codeTabSpaces,
+                                                        fontFamily: encodeURI(window.siyuan.config.editor.fontFamily),
+                                                        workspace: window.siyuan.config.system.workspaceDir,
+
+                                                        title: inbox.shorthandTitle,
+                                                        describe: inbox.shorthandDesc,
+                                                    },
+                                                    config.theme.window.open.windowParams,
+                                                    undefined,
+                                                    undefined,
+                                                    undefined,
+                                                    async (win) => {
+                                                        // 窗口关闭时删除临时文件
+                                                        // console.log(win);
+                                                        setTimeout(async () => rm(temp_file_path_absolute), 0);
+                                                    },
+                                                );
+                                            }
+                                        });
                                     },
                                 ), 0);
                             }